Moves things around a bit
