+++ src/components/price/index.tsx
@@ -1,2 +1,1 @@
 export { Price } from './Price'
-export { defaultConfiguration } from './configuration'
